Put skaffold classes in skaffold package
