+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/JibPlugin.java
@@ -17,6 +17,10 @@
 package com.google.cloud.tools.jib.gradle;
 
 import com.google.cloud.tools.jib.ProjectInfo;
+import com.google.cloud.tools.jib.gradle.skaffold.CheckJibVersionTask;
+import com.google.cloud.tools.jib.gradle.skaffold.FilesTask;
+import com.google.cloud.tools.jib.gradle.skaffold.FilesTaskV2;
+import com.google.cloud.tools.jib.gradle.skaffold.InitTask;
 import com.google.cloud.tools.jib.plugins.common.VersionChecker;
 import com.google.common.annotations.VisibleForTesting;
 import java.util.ArrayList;
@@ -39,19 +43,17 @@ public class JibPlugin implements Plugin<Project> {
 
   @VisibleForTesting static final GradleVersion GRADLE_MIN_VERSION = GradleVersion.version("4.9");
 
-  @VisibleForTesting static final String JIB_EXTENSION_NAME = "jib";
-  @VisibleForTesting static final String BUILD_IMAGE_TASK_NAME = "jib";
-  @VisibleForTesting static final String BUILD_TAR_TASK_NAME = "jibBuildTar";
-  @VisibleForTesting static final String BUILD_DOCKER_TASK_NAME = "jibDockerBuild";
-  @VisibleForTesting static final String FILES_TASK_NAME = "_jibSkaffoldFiles";
-  @VisibleForTesting static final String FILES_TASK_V2_NAME = "_jibSkaffoldFilesV2";
-  @VisibleForTesting static final String INIT_TASK_NAME = "_jibSkaffoldInit";
+  public static final String JIB_EXTENSION_NAME = "jib";
+  public static final String BUILD_IMAGE_TASK_NAME = "jib";
+  public static final String BUILD_TAR_TASK_NAME = "jibBuildTar";
+  public static final String BUILD_DOCKER_TASK_NAME = "jibDockerBuild";
+  public static final String SKAFFOLD_FILES_TASK_NAME = "_jibSkaffoldFiles";
+  public static final String SKAFFOLD_FILES_TASK_V2_NAME = "_jibSkaffoldFilesV2";
+  public static final String SKAFFOLD_INIT_TASK_NAME = "_jibSkaffoldInit";
+  public static final String SKAFFOLD_CHECK_REQUIRED_VERSION_TASK_NAME =
+      "_skaffoldFailIfJibOutOfDate";
 
-  @VisibleForTesting static final String EXPLODED_WAR_TASK_NAME = "jibExplodedWar";
-
-  static final String CHECK_REQUIRED_VERSION_TASK_NAME = "_skaffoldFailIfJibOutOfDate";
-
-  static final String REQUIRED_VERSION_PROPERTY_NAME = "jib.requiredVersion";
+  public static final String REQUIRED_VERSION_PROPERTY_NAME = "jib.requiredVersion";
 
   /**
    * Collects all project dependencies of the style "compile project(':mylib')" for any kind of
@@ -148,18 +150,18 @@ public class JibPlugin implements Plugin<Project> {
             });
 
     tasks
-        .register(FILES_TASK_NAME, FilesTask.class)
+        .register(SKAFFOLD_FILES_TASK_NAME, FilesTask.class)
         .configure(task -> task.setJibExtension(jibExtension));
     tasks
-        .register(FILES_TASK_V2_NAME, FilesTaskV2.class)
+        .register(SKAFFOLD_FILES_TASK_V2_NAME, FilesTaskV2.class)
         .configure(task -> task.setJibExtension(jibExtension));
     tasks
-        .register(INIT_TASK_NAME, SkaffoldInitTask.class)
+        .register(SKAFFOLD_INIT_TASK_NAME, InitTask.class)
         .configure(task -> task.setJibExtension(jibExtension));
 
     // A check to catch older versions of Jib.  This can be removed once we are certain people
     // are using Jib 1.3.1 or later.
-    tasks.register(CHECK_REQUIRED_VERSION_TASK_NAME, CheckJibVersionTask.class);
+    tasks.register(SKAFFOLD_CHECK_REQUIRED_VERSION_TASK_NAME, CheckJibVersionTask.class);
 
     project.afterEvaluate(
         projectAfterEvaluation -> {
