+++ jib-gradle-plugin/src/test/java/com/google/cloud/tools/jib/gradle/skaffold/FilesTaskV2Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google LLC.
+ * Copyright 2019 Google LLC.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,25 +14,27 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.gradle;
+package com.google.cloud.tools.jib.gradle.skaffold;
 
-import com.google.common.base.Splitter;
+import com.google.cloud.tools.jib.gradle.JibPlugin;
+import com.google.cloud.tools.jib.gradle.TestProject;
+import com.google.cloud.tools.jib.plugins.common.SkaffoldFilesOutput;
 import com.google.common.collect.ImmutableList;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.BuildTask;
 import org.gradle.testkit.runner.TaskOutcome;
+import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Test;
 
-/** Tests for {@link FilesTask}. */
-public class FilesTaskTest {
+/** Tests for {@link FilesTaskV2}. */
+public class FilesTaskV2Test {
 
   @ClassRule public static final TestProject simpleTestProject = new TestProject("simple");
 
@@ -43,22 +45,20 @@ public class FilesTaskTest {
    *
    * @param project the project to run the task on
    * @param moduleName the name of the sub-project, or {@code null} if no sub-project
-   * @return the list of paths printed by the task
+   * @return the JSON string printed by the task
    */
-  private static List<Path> verifyTaskSuccess(TestProject project, @Nullable String moduleName) {
+  private static String verifyTaskSuccess(TestProject project, @Nullable String moduleName) {
     String taskName =
-        ":" + (moduleName == null ? "" : moduleName + ":") + JibPlugin.FILES_TASK_NAME;
+        ":" + (moduleName == null ? "" : moduleName + ":") + JibPlugin.SKAFFOLD_FILES_TASK_V2_NAME;
     BuildResult buildResult = project.build(taskName, "-q");
     BuildTask jibTask = buildResult.task(taskName);
     Assert.assertNotNull(jibTask);
     Assert.assertEquals(TaskOutcome.SUCCESS, jibTask.getOutcome());
+    String output = buildResult.getOutput().trim();
+    Assert.assertThat(output, CoreMatchers.startsWith("BEGIN JIB JSON"));
 
-    return Splitter.on(System.lineSeparator())
-        .omitEmptyStrings()
-        .splitToList(buildResult.getOutput())
-        .stream()
-        .map(Paths::get)
-        .collect(Collectors.toList());
+    // Return task output with header removed
+    return output.replace("BEGIN JIB JSON", "").trim();
   }
 
   /**
@@ -69,40 +69,47 @@ public class FilesTaskTest {
    * @param actual the actual list of paths
    * @throws IOException if checking if two files are the same fails
    */
-  private static void assertPathListsAreEqual(List<Path> expected, List<Path> actual)
+  private static void assertPathListsAreEqual(List<Path> expected, List<String> actual)
       throws IOException {
     Assert.assertEquals(expected.size(), actual.size());
     for (int index = 0; index < expected.size(); index++) {
-      Assert.assertEquals(expected.get(index).toRealPath(), actual.get(index).toRealPath());
+      Assert.assertEquals(
+          expected.get(index).toRealPath(), Paths.get(actual.get(index)).toRealPath());
     }
   }
 
   @Test
   public void testFilesTask_singleProject() throws IOException {
     Path projectRoot = simpleTestProject.getProjectRoot();
-    List<Path> result = verifyTaskSuccess(simpleTestProject, null);
-    List<Path> expected =
+    SkaffoldFilesOutput result =
+        new SkaffoldFilesOutput(verifyTaskSuccess(simpleTestProject, null));
+    assertPathListsAreEqual(
+        ImmutableList.of(projectRoot.resolve("build.gradle")), result.getBuild());
+    assertPathListsAreEqual(
         ImmutableList.of(
-            projectRoot.resolve("build.gradle"),
             projectRoot.resolve("src/main/resources"),
             projectRoot.resolve("src/main/java"),
-            projectRoot.resolve("src/main/custom-extra-dir"));
-    assertPathListsAreEqual(expected, result);
+            projectRoot.resolve("src/main/custom-extra-dir")),
+        result.getInputs());
+    Assert.assertEquals(result.getIgnore().size(), 0);
   }
 
   @Test
   public void testFilesTask_multiProjectSimpleService() throws IOException {
     Path projectRoot = multiTestProject.getProjectRoot();
     Path simpleServiceRoot = projectRoot.resolve("simple-service");
-    List<Path> result = verifyTaskSuccess(multiTestProject, "simple-service");
-    List<Path> expected =
+    SkaffoldFilesOutput result =
+        new SkaffoldFilesOutput(verifyTaskSuccess(multiTestProject, "simple-service"));
+    assertPathListsAreEqual(
         ImmutableList.of(
             projectRoot.resolve("build.gradle"),
             projectRoot.resolve("settings.gradle"),
             projectRoot.resolve("gradle.properties"),
-            simpleServiceRoot.resolve("build.gradle"),
-            simpleServiceRoot.resolve("src/main/java"));
-    assertPathListsAreEqual(expected, result);
+            simpleServiceRoot.resolve("build.gradle")),
+        result.getBuild());
+    assertPathListsAreEqual(
+        ImmutableList.of(simpleServiceRoot.resolve("src/main/java")), result.getInputs());
+    Assert.assertEquals(result.getIgnore().size(), 0);
   }
 
   @Test
@@ -110,22 +117,27 @@ public class FilesTaskTest {
     Path projectRoot = multiTestProject.getProjectRoot();
     Path complexServiceRoot = projectRoot.resolve("complex-service");
     Path libRoot = projectRoot.resolve("lib");
-    List<Path> result = verifyTaskSuccess(multiTestProject, "complex-service");
-    List<Path> expected =
+    SkaffoldFilesOutput result =
+        new SkaffoldFilesOutput(verifyTaskSuccess(multiTestProject, "complex-service"));
+    assertPathListsAreEqual(
         ImmutableList.of(
             projectRoot.resolve("build.gradle"),
             projectRoot.resolve("settings.gradle"),
             projectRoot.resolve("gradle.properties"),
             complexServiceRoot.resolve("build.gradle"),
+            libRoot.resolve("build.gradle")),
+        result.getBuild());
+    assertPathListsAreEqual(
+        ImmutableList.of(
             complexServiceRoot.resolve("src/main/extra-resources-1"),
             complexServiceRoot.resolve("src/main/extra-resources-2"),
             complexServiceRoot.resolve("src/main/java"),
             complexServiceRoot.resolve("src/main/other-jib"),
-            libRoot.resolve("build.gradle"),
             libRoot.resolve("src/main/resources"),
             libRoot.resolve("src/main/java"),
             complexServiceRoot.resolve(
-                "local-m2-repo/com/google/cloud/tools/tiny-test-lib/0.0.1-SNAPSHOT/tiny-test-lib-0.0.1-SNAPSHOT.jar"));
-    assertPathListsAreEqual(expected, result);
+                "local-m2-repo/com/google/cloud/tools/tiny-test-lib/0.0.1-SNAPSHOT/tiny-test-lib-0.0.1-SNAPSHOT.jar")),
+        result.getInputs());
+    Assert.assertEquals(result.getIgnore().size(), 0);
   }
 }
